+++ robocorp-code/codegen/commands.py
@@ -574,17 +574,10 @@ COMMANDS = [
     ),
     Command(
         "robocorp.convertProject",
-        "Convert third party RPA file to Robocorp Robot",
+        "Conversion Accelerator from third party RPA to Robocorp Robot",
         add_to_package_json=True,
         server_handled=False,
         hide_from_command_palette=False,
-    ),
-    Command(
-        "robocorp.saveConvertedProject.internal",
-        "Save converted Robocorp Project",
-        add_to_package_json=True,
-        server_handled=True,
-        hide_from_command_palette=True,
     ),
 ]