+++ robocorp-code/vscode-client/src/files.ts
@@ -4,6 +4,7 @@ import * as path from "path";
 import * as fs from "fs";
 import { Uri, window, workspace } from "vscode";
 import { OUTPUT_CHANNEL } from "./channel";
+import { join } from "path";
 
 /**
  * @param mustExist if true, if the returned file does NOT exist, returns undefined.
@@ -56,4 +57,23 @@ export async function readFromFile(targetFile: string) {
 
 export async function writeToFile(targetFile: string, content: string) {
     return await fs.promises.writeFile(targetFile, content);
+}
+
+export async function makeDirs(targetDir: string) {
+    await fs.promises.mkdir(targetDir, { recursive: true });
+}
+
+export async function findNextBasenameIn(folder: string, prefix: string) {
+    const check = join(folder, prefix);
+    if (!(await fileExists(check))) {
+        return prefix; // Use as is directly
+    }
+    for (let i = 1; i < 9999; i++) {
+        const basename = `${prefix}-${i}`;
+        const check = join(folder, basename);
+        if (!(await fileExists(check))) {
+            return basename;
+        }
+    }
+    throw new Error(`Unable to find valid name in ${folder} for prefix: ${prefix}.`);
 }