+++ robocorp-code/vscode-client/src/protocols.ts
@@ -1,24 +1,24 @@
-interface LocalRobotMetadataInfo {
+export interface LocalRobotMetadataInfo {
     name: string;
     directory: string;
     filePath: string;
     yamlContents: object;
 }
 
-interface IVaultInfo {
+export interface IVaultInfo {
     workspaceId: string;
     organizationName: string;
     workspaceName: string;
 }
 
-interface WorkspaceInfo {
+export interface WorkspaceInfo {
     organizationName: string;
     workspaceName: string;
     workspaceId: string;
     packages: PackageInfo[];
 }
 
-interface PackageInfo {
+export interface PackageInfo {
     workspaceId: string;
     workspaceName: string;
     id: string;
@@ -26,40 +26,40 @@ interface PackageInfo {
     sortKey: string;
 }
 
-interface IAccountInfo {
+export interface IAccountInfo {
     fullname: string;
     email: string;
 }
 
-interface ActionResult<T> {
+export interface ActionResult<T> {
     success: boolean;
     message: string;
     result: T;
 }
 
-interface InterpreterInfo {
+export interface InterpreterInfo {
     pythonExe: string;
     environ?: { [key: string]: string };
     additionalPythonpathEntries: string[];
 }
 
-interface ListWorkspacesActionResult {
+export interface ListWorkspacesActionResult {
     success: boolean;
     message: string;
     result: WorkspaceInfo[];
 }
 
-interface RobotTemplate {
+export interface RobotTemplate {
     name: string;
     description: string;
 }
 
-interface WorkItem {
+export interface WorkItem {
     name: string;
     json_path: string;
 }
 
-interface WorkItemsInfo {
+export interface WorkItemsInfo {
     robot_yaml: string; // Full path to the robot which has these work item info
 
     // Full path to the place where input work items are located
@@ -74,18 +74,18 @@ interface WorkItemsInfo {
     new_output_workitem_path: string;
 }
 
-interface ActionResultWorkItems {
+export interface ActionResultWorkItems {
     success: boolean;
     message: string;
     result?: WorkItemsInfo;
 }
 
-interface LibraryVersionDict {
+export interface LibraryVersionDict {
     library: string;
     version: string;
 }
 
-interface LibraryVersionInfoDict {
+export interface LibraryVersionInfoDict {
     success: boolean;
 
     // if success == False, this can be some message to show to the user
@@ -98,20 +98,126 @@ interface LibraryVersionInfoDict {
 
 // these declarations are a superficial variant of the implemented ones in the converter bundle
 // they might need changes if the Converter API is changed
-enum ConversionResultType {
+export enum ConversionResultType {
     SUCCESS = "Success",
     FAILURE = "Failure",
 }
-interface RobotFile {
+
+export interface File {
     content: string;
     filename: string;
 }
-interface ConversionSuccess {
+
+export interface ConversionSuccess {
     type: ConversionResultType.SUCCESS;
-    files: Array<RobotFile>;
+    /**
+     * @deprecated use files
+     */
+    robotFileContent: string;
+    files: Array<File>;
+    report?: File;
+    images?: Array<File>;
+    outputDir: string; // Used internally in Robocorp Code
 }
-interface ConversionFailure {
+
+export interface ConversionFailure {
     type: ConversionResultType.FAILURE;
     error: string;
+    report?: File;
+    images?: Array<File>;
+    outputDir: string; // Used internally in Robocorp Code
 }
-type ConversionResult = ConversionSuccess | ConversionFailure;
+
+export type ConversionResult = ConversionSuccess | ConversionFailure;
+
+export function isSuccessful(result: ConversionResult): result is ConversionSuccess {
+    return result.type === ConversionResultType.SUCCESS;
+}
+
+export enum Format {
+    BLUEPRISM = "blueprism",
+    A360 = "a360",
+    UIPATH = "uipath",
+    AAV11 = "aav11",
+}
+
+export enum ValidationStatus {
+    ValidationSuccess = "ValidationSuccess",
+    ValidationError = "ValidationError",
+}
+
+export interface ValidationSuccess<T> {
+    status: ValidationStatus.ValidationSuccess;
+    payload: T;
+}
+
+export interface ValidationError {
+    status: ValidationStatus.ValidationError;
+    messages: string[];
+    stack?: string;
+}
+
+export type ValidationResult<T> = ValidationSuccess<T> | ValidationError;
+
+export interface Options {
+    objectImplFile?: string;
+    projectFolderPath?: string;
+}
+
+export interface Progress {
+    (amount: number, message: string): void;
+}
+
+export enum CommandType {
+    Analyse = "Analyse",
+    Convert = "Convert",
+    Generate = "Generate",
+}
+
+export interface A360ConvertCommand {
+    command: CommandType.Convert;
+    vendor: Format.A360;
+    projectFolderPath: string;
+    onProgress: Progress;
+    outputRelativePath: string; // Used internally in Robocorp Code
+}
+
+export interface UiPathConvertCommand {
+    command: CommandType.Convert;
+    vendor: Format.UIPATH;
+    projectFolderPath: string;
+    onProgress: Progress;
+    outputRelativePath: string; // Used internally in Robocorp Code
+}
+
+export interface BlueprismConvertCommand {
+    command: CommandType.Convert;
+    vendor: Format.BLUEPRISM;
+    releaseFileContent: string;
+    apiImplementationFolderPath?: string;
+    onProgress: Progress;
+    outputRelativePath: string; // Used internally in Robocorp Code
+}
+
+export interface AAV11GenerateCommand {
+    command: CommandType.Generate;
+    vendor: Format.AAV11;
+    folders: Array<string>;
+    onProgress: Progress;
+    outputRelativePath: string; // Used internally in Robocorp Code
+}
+
+export interface AAV11AnalyseCommand {
+    command: CommandType.Analyse;
+    vendor: Format.AAV11;
+    folders: Array<string>;
+    onProgress: Progress;
+    outputRelativePath: string; // Used internally in Robocorp Code
+}
+
+export type BlueprismCommand = BlueprismConvertCommand;
+export type UiPathCommand = UiPathConvertCommand;
+export type A360Command = A360ConvertCommand;
+export type AAV11Command = AAV11GenerateCommand | AAV11AnalyseCommand;
+
+export type RPAConversionCommand = BlueprismCommand | UiPathCommand | A360Command | AAV11Command;